Handle non-NOR style empty data for EECACHE hostfw data section
Previously the EECACHE data was stored on a PNOR chip. On
PNOR chips the data is intialized to all 0xFF's rather than all 0x00's.
For eBMC systems this data will be stored in the flash file system
on the eBMC. With this change, if we have the CONFIG_PNORDD_IS_PLDM
flag set then we will expect an empty EECACHE image to have all 0x00's
instead of expecting it to be all 0xFF's.

Change-Id: I0b9c0702b5fe26caffc7f994e2213fccde313a10
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114140
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>