Handle modified PHYP HRMOR in multinode dump
There is a scenario where PHYP can be relocated during runtime.
This requires Hostboot to reset the value of ATTR_PAYLOAD_BASE
that is used to identify the start of PHYP's memory.  Only the
primary node is aware of the change, so all of the other hostboot
instances need to ask for the value before processing any of
the dump information.

Change-Id: I58d6acc96e1abe0ce38cce9cdd45e8a21396203d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114248
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>