Update the SBE and reconfig when there is a PG/target state mismatch
If there is a PG/target state mismatch then an SBE update with a reconfig
is necessary to rectify the issue.  As the code is now, the IPL gets
stuck in an infinite loop when only reconfiging without an SBE update.

The updateProcessorSbeSeeproms method is merely the code snippet, from
lines 789 to 854, encapsulated in a method for multiple use purposes.
The code is verbatim, only changing variable names when appropriate,
example l_StepError becomes io_stepErr to represent the new context
in which it is being used.

Change-Id: I47e750e4f54b563a0f7c448b918ff99f24181937
CQ: SW521028
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114165
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>