Added better positive/negative feedback from applying VPD algorithms
Removed the suppression of the error messages from the git commands.
This is no longer needed now that the VPD directory is being compiled
in a single thread. These were needed to suppress the false negatives
when the VPD directory was being compiled with multi threads. Now only
true negatives are happening therefore no need to suppress the error
messages.

Also allow the git fetch command display it's positive feedback
when retrieving the VPD algorithms as opposed to suppressing them.

Change-Id: Ia3e18612677acbf8e3ad4bb32051b390f69c9431
RTC:263370
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114344
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>