Load pldm_base and mctp modules prior to the pnor module
For eBMC systems we will use PLDM file i/o operations to access pnor.
This being the case, we must ensure that we are ready to perform
PLDM requests prior to loading the pnor module. All of the code
required to issue PLDM requests is in the pldm_base module, which
depends on the mctp module being loaded. This commit ensures
both the mctp and pldm_base modules are loaded prior to loading the
pnor module.

Change-Id: I29dff052df9aeaace8ba801020011a602e781e34
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114465
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>