Disable toTarget memoizer on FSP builds
The toTarget memoizer memorizes the result of any request to
convert an entity path to a target.  On FSP, the memoizer will
cache the results of the system target pointer forever, but
that is a problem because the system pointer changes when a
new primary node is elected.  This commit removes
the memoizer from FSP compiles in order to achieve correct
FSP behavior, at the expense of some performance in this
lookup.

Change-Id: I1af66b55b94c1d8be9a1faa9adb653d6984669b3
CQ: SW515162
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114526
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: ARAVIND T NAIR <aravindnair@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>