PLDM Watchdog: getTID
This commit implements support for the getTID command to the BMC.

The BMC responds with its unique TID (Terminus ID) to the getTID request.
The main aim of sending this command to the BMC, however, is that upon
receipt of this command, the BMC will send Hostboot a SetEventReceiver
command that will indicate that HB should arm its internal watchdog
timer.

Change-Id: I2561e5bf4f772d215535ea3d6ffd58c6e554e08c
RTC: 250776
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114802
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>