Secure Boot: Add check for HW Key hash of all PROCs in istep 10.1
Added a check in method retrieveAndExtendSecondaryMeasurements
to validate the HW Key Hash for all secondary PROCs to the primary
PROC. The HW Key Hash is register 1 of the PCR6.

If any secondary PROC's HW Key Hash does not match the primary then
it is noted in the current variable l_mismatch and the logic that is
currently there will create error because of the mismatch.

This check is necessary to ensure the security of the system and to
validate that no SBE has been tampered with.

Also updated a few trace messages to be more explicit of what a
register is.  Example: register 1 is the HW Key Hash.

Change-Id: Ia69987236b8e8263483b0fab6f950a2c13747f16
RTC: 270792
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114660
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>