Secure Boot: Support scratch register-driven TPM SPI Settings
This commit looks for overrides of the TPM SPI divider and receive delay
settings in Scratch Register 13.  It uses a bit in Scratch Register 8 to
see if the values in Scratch Register 13 are valid and should be used.

If the values in Scratch Register 13 are valid:
- The hostboot bootloader code will use the overrides when talking
to the TPM.
- The hostboot code will set the overrides into ATTR_TPM_SPI_BUS_DIV
so that they can be used by the TPM SPI code and also passed on to
the secondary SBEs.

If not valid, the default settings will continue to be used.

Change-Id: Ie3442ba68ff3f9d7b38bb60f3604748e85c4fa4e
RTC:270358
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118157
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>