Enhance resiliency of GARD tests
During cxxtests its possible that a previous gard record can exist.

The HWAS tests are currently run in the serial test bucket, however this
does not entirely protect the HWAS tests that are sensitive to any
possible gard records that may be previously introduced for expected
or unexpected reasons.

To proactively sense and clear any non-related gard records prior to the
function of the HWAS tests, code is added to detect and cleanup
as needed in areas where previous influence is undesirable to the HWAS
test outcome.

The gard records will be dumped out where the EID of the gard
record previously created can be identified to help root cause any other
tests that may need attention and cleanup.  The dump also includes
a binary dump to aide in analysis.

Change-Id: Ice3d33e664c4113db336dfa2dc744057a6d93f19
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114823
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>