Add security checks for measurement seeprom
Adds checks to ensure that the minimum secure version of the measurement
code is written to the measurement SEEPROMs and that the
OTP_SPIM_MEAS_SEEPROM_LOCK fuse has been blown. Once that fuse is blown
the measurement version is locked in and can't be changed as that is a
hardware fuse.

Change-Id: I5532c9e2aea5dbb9eae620e061ca78eb4ff10480
RTC:250781
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115257
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>