PRD: Invalid write to EQ_CHIPLET_RE_FIR
We're clearing EQ_CHIPLET_RE_FIR[2] during post analysis of a core unit
checkstop attention. This register is not writable in P10. The hardware
team says we should not need to manually clear that bit in P10 because
it should be level driven and will clear itself when the chiplet FIRs
for the core are masked at the end of analysis.

Change-Id: If7c709a3a015ede9e214d50cedd047c24ddfaf54
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115158
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>