Enhancements to i2c reset logic
We have encountered cases where the i2c logic on the P10 is getting
into a state that our current Hostboot code cannot recover from.
This is likely due to some known issues on the DD1 hardware, but
regardless it has to be handled.  Previously we did a simple
reset of the engine registers, but that is now expanded to do
more of what the FSP has been doing (which seemed to make things
work better for us).

The complete reset sequence is now
- clear interrupt masks
- reset port busy
- send a stop
- reset the engine registers
- reset any sticky errors in the engine (and the fifo)
- reset port busy (again)
- send stop to all devices

Change-Id: Ib35ae80ac5e57109a4cdfd5273db9a893b845cf5
CQ: SW520943
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115099
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>