Move OCMB_FW_UPDATE_STATUS attr from SYSTEM -> NODE target
Right now we use system-level attribute to determine if a
node needs one or more of it's OCMBs updated. When a non-zero
node needs ocmbs updated but node0 does not this becomes a problem
because only one node's system attribute needs to win if they
don't all match. In this case I believe HWSV will take node0's
system attribute values since it is the primary node. This is
causing node3 to never update its ocmbs. Since we are never
updating the downlevel OCMB we end up in an infinite loop.

Change-Id: Ia14a19fb20c3ab71c7940ab0d6169e9d44a12246
CQ: SW522642
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115107
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>