PRD: issues with neighbor core attns in EQ_CORE_FIR
The current design for neighbor core attentions (EQ_CORE_FIR[56:57]) was
to let the rule code handle the isolation and to simply ignore these
attentions. The problem is that post analysis will kick in after
analyzing any core, which will mask attentions on both cores for a unit
checkstop. This would cause issues when we go to analyze the other core
because the attentions were masked.

Change-Id: Ibf507d5e454e84b6a121b418a3e0431637a31875
CQ: SW522001
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115154
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>