PLDM Watchdog: Terminus Locator PDR
This commit creates a Terminus Locator PDR that HB passes to BMC
along with other PDRs. The Terminus Locator PDR serves to let
BMC know whom to monitor for watchdogs. The PDR is set to valid
at the point of creation, and when HB transitions to payload,
it invalidates the PDR to signal the BMC that it should stop
watching HB for watchdogs and start watching the payload.

Change-Id: Idec51b74aaa96d79a3e272a24a25f6e59de1503f
RTC: 250776
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115190
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>