Enable i2c device callouts for FSP systems
When encountering I2C failures, it is typical that the underlying
cause is the end device.  However, the current Hostboot code is
only looking for a small subset of possible devices for callouts
when running on FSP systems.  This change will enable the logic
that is currently running on openpower systems to get a callout
of end device along with the i2c master (processor).

Note - This solution is still not ideal as we don't have the
full bus path covered by callouts, but it is substantially better
than what currently exists.

Change-Id: Ic0fbc5f66f9900ed396aeb70e9147626545a7704
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115214
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>