Move ATTR_BLOCK_SPEC_DECONFIG to node level
On a multinode system there is only one master copy of the
system level attributes stored on the FSP.  However, every hostboot
instance will have its own version while it is running.  If each
instance were to put a different value for a writable system-level
nonvolatile attribute, that value will not be preserved to the next
IPL.  Instead, only the value from the primary node will be used.

ATTR_BLOCK_SPEC_DECONFIG is one such attribute.  Losing the data
in this attribute could cause unexpected behavior on a multinode
system.

Another deprecated attribute was also removed while reviewing the
xml.

Change-Id: Ib0995790bab6ceb0cf0c2906c7d14071f71f1c70
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115300
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>