Pull in changes for mctp_msg_ctx_add_pkt function from upstream
There was some optimizations done to this function in the upstream
project that we want. There are plans down the road to pull in the
upstream as a subtree similar to what we are doing for the pldm
subtree. These upstream changes introduce a maximum message size
for MCTP messages. This protects the core logic from allocating
more than X bytes for a given message context. The default that
the bmc is using is 64KB. Hostboot makes PLDM requests for up
to 128KB reads so we will set the limit to the larest PLDM message
we can expect.

Change-Id: Ic562cad76b71212eb7b788e4720374b898d791e7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115322
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>