Only allocate as much memory that is needed for mctp packets
Hostboot hit memory fragmentation issues when we enabled
PNOR I/O via PLDM. This was because the function
mctp_pktbuf_alloc was always allocating an MTU (32KB) of data
for every PLDM request, even if the request was only 20
bytes. Due to the way memory is allocated by Hostboot's kernel,
prior to exiting cache containment, it difficult for Hostboot to come up
with 8 consecutive pages without pinning the memory. As a workaround
we will only allocate as much space as is needed for a given
packet. This will limit libmctp's ability to add to this
packet later, which is a feature that Hostboot does not
currently use. It is not clear if this change can be
upstreamed and this issue may be revisted in the future.

Change-Id: I205d92bff63af31f85d815ebee79ba99f23246e2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115323
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>