Add hook to call HWP when HBRT starts
Added a generic path to execute arbitrary hardware procedures as
part of HBRT start/restart in order to apply any new settings on
a concurrent code update.

Add explicit support for new p9_fbc_tl_disable_gathering HWP to
run in this path.  Note that it is okay for it to run on all
starts, the data will just be rewritten to the same value.

Change-Id: Ie5e09b353bf7a685f2b2d49ae88ff270eebda55c
CQ: SW523479
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115444
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>