OCMB I2C Security Configuration
P10 DD2 implements new registers to dynamically configure
the I2C devAddrs on which to block OCMB I2C reads and writes.

In coordination with the SEEPROM UPDATE LOCK (SULbit) and
the Secure OCMB Lock (SOLbit) the desired OCMB I2C devAddrs
will be written to the proper device protection registers
to configure the Engine's devAddr which is desired to be
protected from reads and writes.

Engine A device protection register is configured prior
to the setting of the SULbit.  Once the SULbit is set Engine
A's device protection register is no longer able to be updated.
Currently this logic is placed in host_secureboot_lockdown.

Engine B, C and E device protection registers are configured
after the SULbit is set but prior to setting the SOLbit.
Once the SOLbit is set the device protection registers are no
longer able to be updated.  Currently this logic is placed in
host_secure_rng (which is after memory initialization is complete).

After the system is powered on, the CBS start will reset all
the device protection registers to "open".

After the SULbit, SOLbit and device protection registers
are configured and locked down the Host will only use
inband access to OMI.

Change-Id: Ieff8b5bb5211e2e93dcf99d00991dedb9260af08
RTC:266776
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115493
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>