Force TPM present and guarded on TPM presence mismatch
If target could have a deferred presence detect, and a
presence mismatch is detected between FSP and Hostboot,
fatally guard it to prevent Hostboot from inadvertently
resurrecting it later and triggering endless failovers.

One such case, absent this mitigation, is if
both TPMs in a node hit a mismatch where FSP detects
TPM as present but Hostboot does not, and the
"TPM Required Policy" is "Required".  In that case
the system flags the primary TPM as not present
then fails over to the other FSP and attempts a new
boot.  On that boot, discovery assumes the backup TPM
is present and functional until it can check it in istep
10.3, but the bad primary TPM forces a termination first,
causing the backup's "good" present/functional status to
replicate down to the FSP.  This fools the alignment check
to think the original primary TPM is now good and thus
alignment check initiates a fail over back to the original
FSP where the cycle starts over.

Guard records can only be placed on present targets, so
on a TPM presence mismatch this forces the TPM to present
before guarding it.

Change-Id: I3191becf3fad6fd236075d18602422e3078eb203
CQ: SW522858
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115834
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>