Increase HBRT base max transmission unit to from 252 bytes -> 4092
When HBRT's MCTP stack was initially implemented the MTU was fixed
at 252 bytes. Since then a new version of MCTP was added to allow
PHYP and the BMC to negotiate the MTU. PHYP and the BMC are currently
negotiating a 4KB-4 MTU so we will update our BMTU to 4KB-4 to compensate.
A path still needs to be added for PHYP to inform us what the
negotiated MTU is as HBRT does not have access to the LPC address
space where this information is stored. This has not been a problem
up until this point because all PLDM messages used at runtime have
been less than 252 bytes. This however, will be a problem when we
attempt to read in hostfw lids at runtime.

Change-Id: I4c649b0978f8461f273a94487279fe5acee4d900
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115882
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>