Create/Move Attributes to Node Level for Key Clear and Physical Presence
This commit both creates and moves attributes related to Key Clear Requests
and Physical Presence Detection to the node level.  This is to fix a bug
where a node that requires a re-IPL might clear out the key clear request.

The FSP/HB shared ATTR_KEY_CLEAR_REQUEST will remain at system level, but
is has also been added at the node level.  The FSP code will be responsible
for setting the node-level attributes based on the system level one.

The following HB-only attributes were moved from the system level to the
node level:
- ATTR_KEY_CLEAR_REQUEST_HB
- ATTR_PHYS_PRES_REQUEST_OPEN_WINDOW
- ATTR_PHYS_PRES_FAKE_ASSERT
- ATTR_PHYS_PRES_ASSERTED
- ATTR_PHYS_PRES_REIPL

Also, the clearing of the ATTR_KEY_CLEAR_REQUEST at the node level has been
moved to istep 16 before the attributes are synced down to the FSP.  There was
also a minor testcase updated related to moving an attribute to the node level.

Change-Id: I26dbebd7b425267afc20a44a99b3c294a3dd7bba
CQ:SW523488
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115962
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>