Add param to add review to commits generated by update_pldm_subtree.sh
This commit updates update_pldm_subtree.sh so it will setup the hostboot
environment automatically and allow the caller to specify an email to
add as a reviewer to the gerrit commit that gets generated from this
script. This will make less work for the CI team who will setup a job
to call this.

Change-Id: I30bec065d12209cf6090056eef220c17a7c8709d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116112
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>