Handle Special Case of Minimum Secure Version
This commit looks for and handles a special case in regards to
the system's Minimum Secure Version (MSV):
- If the MSV lockin policy is false --AND--
- The incoming MSV value is >= 1 --AND--
- The current Minimum Secure Version is zero -- AND --
- Secureboot is enabled...
Then set the minimum secure version to 1.

This is considered a special case because the MSV lockin policy is
being overruled and the value of 1 is not necessarily the exact
incoming MSV value.

This change is necessary for early ship customers who might get a
driver with SV==0, but we want to automatically update them to at
least 1 once they get an official production build without
making them use the ASMI "Lockin Policy" option.

Change-Id: Id8476c19c875c5660fc263c94614c2bbe06fdfcb
RTC:268451
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116057
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>