Commit PRD Errors In PM Reset Flow
This commit introduces change to commit any error logs produced by
the call to PRD attn handler in PM reset flow. This is done so
that the PM complex flow (that calls PM reset) can continue despite
the PRD error. Whether we need to still call the attn handler in
that flow in P10 is still to be discussed.

Change-Id: Id1183be3644a0cad20667c9797ac4bf0728c976f
CQ: SW524190
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116081
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>