Default to include the backtraces for the PS tool
This is intended to be a quality of life improvement for developers.
In a majority of use-cases developers are passing this flag. To
save some extra typing this will default the --with-backtrace flag
to being on. This commit also introduces --without-backtrace which
will hide the backtraces to provide a less resource intensive option.
Passing the --with-backtrace flag will not raise an errors it is
just now ignored.

Change-Id: If249f54290f15877646c674fe39734019315c8e0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116115
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>