PRD: remove PM recovery support
PM recovery design changed such that it no longer needs to set a FIR bit
and have PRD react to the attention. That design was intended for P8
when HBRT need to communicate with PRD on the FSP. Instead, the PM
recovery code will cut PRD out as the middle man and call the PM callout
HWP itself to ensure that is done before exiting the process.

Change-Id: I9bf76bca535826e7b9a4e37a1841b5b0c047e5f2
CQ: SW524190
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116269
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>