Force BAR recompute after topology id swap
In the case where there is no functional memory behind proc0 on
a node, we swap the effective topology id of proc0 with a proc
that has memory in order to always maintain a physical address
zero.  If one of the swapped procs is the boot processor, we
handle this early in the boot and force a reconfig loop to
reboot the SBE with the new topology id value.  However, if we
are swapping only with the secondary processors we won't trigger
a reconfig loop.  This means that even though we have correct
topology ids setup, we don't fix all of the other downstream
attributes that are based on that.

The fix is to explicitly rewrite all of the BAR attributes (the
same ones as in the existing logic) at the point where we do
any kind of topology id swap.  We also need to force a reconfig
loop so that we rerun the memory configuration steps that
execute before we do the swap.

Change-Id: I3c2a83e41d9726e3d66cf847c7539fba37a7bbcb
CQ: SW523085
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116334
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>